Tryout for on demand switching between BLE and Wifi counting in libpax
