+++ include/libpax_helpers.h
@@ -1,0 +1,10 @@
+#ifndef _LIBPAX_HELPERS_H
+#define _LIBPAX_HELPERS_H
+
+#include "globals.h"
+#include <stdio.h>
+#include <libpax_api.h>
+
+void init_libpax();
+
+#endif
